backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-Cors==4.0.1
Flasgger==0.9.7.1
PyMySQL==1.1.1
PyPDF2==3.0.1
scikit-learn==1.5.1
python-dotenv==1.0.1
Werkzeug==3.0.3